+++ src/models/command/blockAutomaticCommand.ts
@@ -3,19 +3,19 @@ import { CommandSource } from './commandSource';
 import { CollisionSolving } from '../collisionSolving';
 import { BaseCommand } from './baseCommand';
 
-export class AutomaticBlockDisableUntilCommand extends BaseCommand {
-  public _commandType: CommandType = CommandType.AutomaticBlockDisableUntilCommand;
+export class BlockAutomaticCommand extends BaseCommand {
+  public _commandType: CommandType = CommandType.BlockAutomaticCommand;
 
   /**
-   * Command to disable automatic actions until a specific date.
+   * Command to disable automatic actions for a specific duration.
    * @param {CommandSource} source The source of the command.
-   * @param {Date} targetDate The date until the automatic actions will be disabled.
+   * @param {number} durationMS The duration in milliseconds for the automatic actions to be disabled.
    * @param reason You can provide an individual reason here for debugging purpose.
    * @param {CollisionSolving} onCollideAction The action to take if a block is already active. --> Default: overrideIfGreater
    */
   public constructor(
     source: CommandSource | BaseCommand,
-    public targetDate: Date,
+    public durationMS: number,
     reason: string = '',
     public onCollideAction: CollisionSolving = CollisionSolving.overrideIfGreater,
   ) {